requirements/testing.txt
# Pytest for running the tests.
pytest>=6.1,<7.0
pytest-cov>=2.10.1,<3.0
pytest-django>=4.1.0,<5.0